requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
PyJWT==1.6.4
requests==2.20.0
urllib3==1.23
